Solve day 4
We're currently iterating over the map a lot. Once to redefine accessibility, then again to figure out which are accessible. We could carry a number of accessible rolls in the grid, as a kind of cache. And we could take removed rolls from the current iteration into account for the accessible rolls around it.

Or recalculate accessibility only for rolls that were next to an accessible location. No need to iterate over the whole map.

Maybe as a next step.