Solve day 5
I spent a lot of time trying to realize this as a form of stream process. Like you'd have with RxJS. Unfortunately, I could not find a way to have a stream process some elements and return the remaining stream. I always ended up processing the first elements multiple times. And that was the step I wanted to only do once.

So in the end, I went back to processing the whole file in memory.

Also, I looked into representing the list of fresh ingredient id ranges as a binary search tree. Any search tree implementations are libraries though. So I ended up not going down that path either.

Simple exercise and still I spent too much time on it :)