Solve day 8 part 2
Keeping track of the last connection and iterating n times where n is the number of potential connections are both not _nice_. But for keeping to solve both problems with the same code, I thought of them as acceptable tradeoffs.

Also, we're currently connecting the first 1000 connections twice. Not ideal, but acceptable I'd say.